pyproject.toml
[tool.poetry]
name = "shawns-nice-tools"
version = "0.1.0"
description = "random nice tooling to make life easier"
authors = ["shawn"]
license = "MIT"
package-mode = false

[tool.poetry.dependencies]
python = "^3.10"
boto3 = "^1.26"
rich = "^12.0"
tomli = "^1.2"
pyyaml = "^6.0"

[tool.poetry.dev-dependencies]
pytest = "^7.0"
pytest-mock = "^3.6"

[tool.poetry.group.dev.dependencies]
black = "^24.10.0"
isort = "^5.13.2"
ruff = "^0.1"
pre-commit = "^4.0.1"
pyflakes = "^3.2.0"
autopep8 = "^2.3.1"
flake8 = "^7.1.1"

[build-system]
requires = ["poetry>=1.0"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
fix = true
line-length = 120
select = ["E", "F", "I"]
ignore = ["E501"]

[tool.black]
line-length = 120

# Pytest configuration section
[tool.pytest.ini_options]
testpaths = ["aws/tests"]
pythonpath = ["aws"]
addopts = "--import-mode=append"
